adding photo grid permissions

+++ app/Http/Livewire/Photo/Remove/RemoveGridPhoto.php
@@ -4,9 +4,11 @@ namespace App\Http\Livewire\Photo\Remove;
 
 
 use App\Models\GalleryImage;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Support\Facades\URL;
 use LivewireUI\Modal\ModalComponent;
 use function view;
@@ -17,6 +19,12 @@ use function view;
  */
 class RemoveGridPhoto extends ModalComponent
 {
+    /**
+     * Provide authorization functionality for permissions
+     * verification.
+     */
+    use AuthorizesRequests;
+
     /**
      * the page_module model reference that will be
      * used as a reference to update the page_modules
@@ -36,9 +44,13 @@ class RemoveGridPhoto extends ModalComponent
      * the function that when called will delete
      * the photo from the gallery
      * @return void
+     * @throws AuthorizationException
      */
     public function delete()
     {
+        // verify authorization
+        $this->authorize('delete photo');
+
         // delete the image
         GalleryImage::where('id', '=', $this->photo_id)->delete();
